refactor: extract FiltersPrice
